+++ benchmarks/workload_test.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/tobilg/golyglot"
+	"github.com/renart-data/golyglot"
 )
 
 const workloadManifestEnvironment = "GOLYGLOT_BENCH_WORKLOAD"